MP Party Affiliation & Constituents Brexit Votes
Literature Review - "Members of Parliament ("MP") Party Affiliation: How it Influenced Constituents' Votes in the EU Referendum"

This project stopped short of doing actual data analysis due to time constraints. It's a topic that is interesting to me and I would like to do further analysis on this; particularly since the UK has officially left the EU and further implications of the Brexit vote are being seen daily.


Abstract:
When UK citizens voted via national referendum to leave the European Union, there was an international reaction. While there have been many studies on the regional and demographic variables that have influenced voter behavior, there are not currently any studies on if Members of Parliament political party affiliation influenced their constituents voting behavior in the EU referendum. This paper looks at if and how MPs political affiliation influenced individuals in the Brexit vote by looking at individual level data that’s already been collected, with a proposal for additional data to be gathered and compared. I argue that constituents were not influenced by their MP’s party affiliation in the Brexit vote and that there is a serious issue with effective MP representation of their constituency.